Article listing and more.

+++ src/db/list_art_query.h
@@ -6,7 +6,7 @@
 namespace db {
 	class ListArtQuery : public Query {
 	public:
-		ListArtQuery(Database &db, int ng_id) : Query(db), newsgroup_id(ng_id) {}
+		ListArtQuery(Database &db, size_t ng_id) : Query(db), newsgroup_id(ng_id) {}
 		Result *execute();
 	private:
 		int newsgroup_id;
